+++ lib/an_viewmodel.dart
@@ -1,5 +1,6 @@
 library an_viewmodel;
 
+export 'src/context_ext.dart';
 export 'src/view_model.dart';
-export 'src/view_model_ref.dart';
 export 'src/view_model_ext.dart';
+export 'src/view_model_ref.dart';